backend/app/config.py
from dotenv import load_dotenv
import os

load_dotenv("../.env")

DATABASE_URL = os.getenv("DATABASE_URL")

JWT_SECRET = os.getenv("JWT_SECRET")
JWT_ALGORITHM = "HS256"

print("Loaded DATABASE_URL =", DATABASE_URL)