backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
scikit-learn
numpy
twilio
python-dotenv
passlib[bcrypt]
python-jose